Feature: Added mRMR R Script to execute Feature Selection
The R mRMR library runs faster than other libraries such as the C/C++ mRMR library found online. Python libraries for mRMR are discontinued and unreliable. So R is the best option

However, we can't have R and Python Kernels running in a notebook at the same time, so we will be running this script through subprocesses