Notebook: Scored PCA Selected Features through Cross Validation
Ran PCA with different Variance Retain (5, 25, 45, 65, 85). This is because the different variance retain resulted in the different scoring. So, instead of scoring with variance retain 80, 85, 90, 95, I went with scoring the variance across many different further apart variances. Accordingly, we will choose which PCA option is best for the different models